% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Fetch all FII's (Brazilian Real Estate Funds)}
\usage{
fetch()
}
\value{
data.frame of all fii's.
}
\description{
Fetch all FII's (Brazilian Real Estate Funds)
}
\examples{
fetch()
}
