% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reinvest.R
\name{reinvest}
\alias{reinvest}
\title{Find the best reinvestment opportunities}
\usage{
reinvest(criteria = "dividend")
}
\arguments{
\item{criteria}{reinvestment criteria. "dividend" or "price"}
}
\value{
assets sorted according to the reinvestment criteria.
}
\description{
\code{reinvest} takes a list of assets stored in \code{assets.yaml} and
returns the best reinvestment opportunities at the moment.
}
\examples{
reinvest()
}
